+++ django_grapesjs/forms/fields.py
@@ -2,6 +2,7 @@ from django import forms
 
 from django_grapesjs.forms import GrapesJsWidget
 from django_grapesjs.settings import BASE, GRAPESJS_DEFAULT_HTML, REDACTOR_CONFIG
+from django_grapesjs.utils import apply_string_handling
 
 __all__ = (
     'GrapesJsField',
@@ -15,11 +16,22 @@ class GrapesJsField(forms.CharField):
 
     widget = GrapesJsWidget
 
-    def __init__(self, max_length=None, min_length=None, strip=True, empty_value='',
-                 default_html=GRAPESJS_DEFAULT_HTML, html_name_init_conf=REDACTOR_CONFIG[BASE],
-                 *args, **kwargs):
-        super().__init__(max_length=None, min_length=None, strip=True, empty_value='', *args, **kwargs)
+    def __init__(self, default_html=GRAPESJS_DEFAULT_HTML, html_name_init_conf=REDACTOR_CONFIG[BASE],
+                 apply_django_tag=False, validate=False, *args, **kwargs):
+        super().__init__(*args, **kwargs)
 
         self.widget.default_html = default_html
         self.widget.html_name_init_conf = html_name_init_conf
+        self.widget.apply_django_tag = apply_django_tag
+        self.validate = validate
+
+    def validate(self, value):
+        super().validate(value)
+        # TODO: check the field
+        # if self.validate:
+
+    def clean(self, value):
+        value = apply_string_handling(value, 'apply_tag_save')
+
+        return super().clean(value)