+++ django_grapesjs/forms/widgets.py
@@ -3,6 +3,7 @@ from django.template.loader import render_to_string
 
 from django_grapesjs.settings import GRAPESJS_TEMPLATE
 from django_grapesjs.template import template_source
+from django_grapesjs.utils import apply_string_handling
 
 __all__ = (
     'GrapesJsWidget',
@@ -44,17 +45,21 @@ class GrapesJsWidget(forms.Textarea):
         context = super().get_context(name, value, attrs)
 
         context['widget']['attrs']['id'] = self.get_formated_value_id(context['widget']['attrs']['id'])
-        context['widget'].update({'get_render_html_value': self.get_render_html_value(self.default_html)})
-        context['widget'].update({'html_name_init_conf': self.html_name_init_conf})
+        context['widget'].update({
+            'get_render_html_value': self.get_render_html_value(),
+            'html_name_init_conf': self.html_name_init_conf
+        })
 
         return context
 
-    def get_render_html_value(self, default_html, apply_django_tag=False):
+    def get_render_html_value(self):
         def _get_render_html_value():
-            if not apply_django_tag:
-                return template_source.get_template(default_html).template
+            if not self.apply_django_tag:
+                return apply_string_handling(
+                    template_source.get_template(self.default_html).template
+                )
 
-            return render_to_string(default_html)
+            return apply_string_handling(render_to_string(self.default_html))
 
         return _get_render_html_value